% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpk.R
\docType{data}
\name{Jones2013}
\alias{Jones2013}
\title{Jones 2013 Human PBPK model}
\format{An object of class \code{RxODE} of length 37.}
\usage{
Jones2013
}
\description{
Jones 2013 Human PBPK model
}
\examples{
library(RxODE)
Jones2013 \%>\% solve(et(amt=100,cmt=D,ii=4,until=24) \%>\% et(seq(0,24,length.out=100))) \%>\% plot

}
\references{
1: Jones H, Rowland-Yeo K. Basic concepts in
    physiologically based pharmacokinetic modeling in drug
    discovery and development. CPT Pharmacometrics Syst
    Pharmacol. 2013 Aug 14;2:e63. doi: 10.1038/psp.2013.41. PubMed
    PMID: 23945604; PubMed Central PMCID: PMC3828005.

https://github.com/metrumresearchgroup/mrgsolve/blob/9ff7240da48ec133635a6f0db2143f13623f0aff/inst/models/pbpk.cpp
}
\keyword{datasets}
