% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk3cmt-doc.R
\docType{data}
\name{pk3cmt}
\alias{pk3cmt}
\title{pk3cmt RxODE model}
\format{An \emph{RxODE} model with 22 parameters, 2 ODE states, and 1 calc vars.

\emph{Parameters (pk3cmt$params)}

\describe{
  \item{popCl}{Population Clearance value (default=1)}
  \item{popV}{Population Volume of distribution (default=20)}
  \item{popKa}{Population Absorption coefficient (default=1)}
  \item{popVp}{Population Peripheral Volume of Distribution #1 (default=10)}
  \item{popVp2}{Population Peripheral Volume of Distribution #2 (default=100)}
  \item{popQ}{Population Inter-Compartmental Clearance of Compartment #1 (default=2)}
  \item{popQ2}{Population Inter-Compartmental Clearance of Compartment #2 (default=2)}
  \item{bsvCl}{Between Subject Variability for Clearance (default=0)}
  \item{bsvV}{Between Subject Variability for Volume (default=0)}
  \item{bsvKa}{Between Subject Variability for Ka (default=0)}
  \item{bsvVp}{Between Subject Variability for Peripheral Volume of Distribution #1 (default=0)}
  \item{bsvVp2}{Between Subject Variability for Peripheral Volume for Compartment #2 (default=0)}
  \item{bsvQ}{Between Subject Variability for Inter-compartmental Clearance #1 (default=0)}
  \item{bsvQ2}{Between Subject Variability for Inter-compartmental Clearance #2 (default=0)}
  \item{popLagDepot}{Population Value for Lagged doses on Central Compartment (default=0)}
  \item{popLagCentral}{Population Value for Lagged doses on Central compartment (default=0)}
  \item{popRateCentral}{Population Value for modeled rates (default=0)}
  \item{popDurCentral}{Population Value for modeled duration (default=0)}
  \item{bsvLagDepot}{Between Subject variability for depot (default=0)}
  \item{bsvLagCentral}{Between Subject Variability for Central Lag-time (default=0)}
  \item{bsvRateCentral}{Between Subject Variability for Central Rate (default=0)}
  \item{bsvDurCentral}{Between Subject Variability for Central Duration (default=0)}
}

\emph{State pk3cmt$state}

\describe{
  \item{depot}{Depot Compartment (=1)}
  \item{central}{Central compartment (=2)}
}

\emph{Calculated Variables pk3cmt$lhs}

\describe{
  \item{cp}{Concentration in the plasma}
}}
\usage{
pk3cmt
}
\description{
pk3cmt RxODE model
}
\examples{
## To see the model code:
summary(pk3cmt)

pk3cmt \%>\% solve(et(timeUnits="hr") \%>\% et(amt=100)) \%>\% plot # Oral
pk3cmt \%>\% solve(et(amt=100,cmt=central,timeUnits="hr")) \%>\% plot # IV
}
\seealso{
\code{\link[RxODE]{eventTable}}, \code{\link[RxODE]{et}}, \code{\link[RxODE]{rxSolve}}, \code{\link[RxODE]{RxODE}}
}
\keyword{datasets}
