% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{RxODE}
\alias{et}
\alias{etRep}
\alias{etSeq}
\alias{as.et}
\alias{eventTable}
\alias{add.dosing}
\alias{add.sampling}
\alias{rxControl}
\alias{rxClean}
\alias{rxUse}
\alias{rxShiny}
\alias{genShinyApp.template}
\alias{cvPost}
\alias{\%>\%}
\alias{rxSolve}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{RxODE}{\code{\link[RxODE]{RxODE}}, \code{\link[RxODE]{et}}, \code{\link[RxODE]{etRep}}, \code{\link[RxODE]{etSeq}}, \code{\link[RxODE]{as.et}}, \code{\link[RxODE]{eventTable}}, \code{\link[RxODE]{add.dosing}}, \code{\link[RxODE]{add.sampling}}, \code{\link[RxODE]{rxControl}}, \code{\link[RxODE]{rxClean}}, \code{\link[RxODE]{rxUse}}, \code{\link[RxODE]{rxShiny}}, \code{\link[RxODE]{genShinyApp.template}}, \code{\link[RxODE]{cvPost}}, \code{\link[RxODE]{\%>\%}}, \code{\link[RxODE]{rxSolve}}}
}}

