% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpk.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{RxODE}
\alias{et}
\alias{add.sampling}
\alias{add.dosing}
\alias{eventTable}
\alias{rxClean}
\alias{rxSolve}
\alias{rxControl}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{RxODE}{\code{\link[RxODE]{RxODE}}, \code{\link[RxODE]{et}}, \code{\link[RxODE]{add.sampling}}, \code{\link[RxODE]{add.dosing}}, \code{\link[RxODE]{eventTable}}, \code{\link[RxODE]{rxClean}}, \code{\link[RxODE]{rxSolve}}, \code{\link[RxODE]{rxControl}}}
}}

